Feat: 백엔드 API수정에 따른 service 파일 일부 수정 / 로그인, 회원가입 API 연동 완료

+++ lib/model/SponModel.dart
@@ -11,7 +11,7 @@ Spon SponJson(String str) => Spon.fromJson(json.decode(str));
 String SponToJson(Spon data) => json.encode(data.toJson());
 
 class Spon {
-  Long spon_id;
+  int spon_id;
   SMember sMember;
   Post post;
   Ingredients ingredients;
